requirements.txt
django~=5.2.5
gunicorn
psycopg2-binary
Pillow~=10.4
whitenoise==6.7.0
django-timezone-field
django-countries==7.6.1
django-image-cropping
easy-thumbnails
django-ckeditor>=6.7.1